Get Vendor Payments While Price is double to get precise number

+++ ViewModels/PaymentViewModel/VendorPaymentViewModel.cs
@@ -10,7 +10,7 @@ namespace ViewModels.PaymentViewModel
 {
     public class VendorPaymentViewModel
     {
-        public required float Price { get; set; }
+        public required double Price { get; set; }
         public required DateTime DateTime { get; set; }
         public required string ServiceTitle { get; set; }
         public required UserMinInfoViewModel User { get; set; }
